% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aff_death_haz}
\alias{aff_death_haz}
\title{Constant hazard of death for affected (diseased) individuals.}
\format{
A numeric value.
}
\source{
Schauer, M., Peiper, M., Theisen, J., and Knoefel, W. (2011). Prognostic factors in
patients with diffuse type gastric cancer (linitis plastica) after operative treatment. European
Journal of Medical Research, 16(1):29–33.
}
\usage{
aff_death_haz
}
\description{
A constant hazard rate of death for individuals affected by the disease that is
linked to the rare variant (e.g. gastric cancer).
}
\keyword{datasets}
