% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dis_haz}
\alias{dis_haz}
\title{Age-specific hazard of disease (e.g. gastric cancer) in the general population.}
\format{
A numeric vector of length 100.
}
\source{
SEER (2025).
SEER*Explorer: An interactive website for SEER cancer statistics
Surveillance Research Program, National Cancer Institute.
https://seer.cancer.gov/statistics-network/explorer/. Data source(s): SEER Incidence
Data, November 2024 Submission (1975-2022)
}
\usage{
dis_haz
}
\description{
A numeric vector containing the age-specific hazard rates of
gastric cancer in the general population. The rates are expressed
per person-year.
}
\keyword{datasets}
