% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree.R
\name{pedigree}
\alias{pedigree}
\title{Single pedigree simulation}
\usage{
pedigree(
  aff,
  frailty,
  frail_var,
  dis_haz,
  unaff_death_haz,
  aff_death_haz,
  a_start,
  a_end,
  f_start,
  f_end,
  GRR
)
}
\arguments{
\item{aff}{Number of disease cases among the carriers required for the pedigree to be ascertained.}

\item{frailty}{Indicator for whether include (1) or not include (0) a frailty term.}

\item{frail_var}{Value of the frailty variance, if frailty included.}

\item{dis_haz}{Hazard of the disease in the population of interest.}

\item{unaff_death_haz}{Hazard of death in the population of interest.}

\item{aff_death_haz}{Hazard of death in affected (diseased) population.}

\item{a_start}{Start year of the ascertainment interval (interval in which the proband develops the disease).}

\item{a_end}{Ending year of the ascertainment interval (interval in which the proband develops the disease).}

\item{f_start}{Start year of the founder birth interval (interval in which the founder is randomly born).}

\item{f_end}{Ending year of the founder birth interval (interval in which the founder is randomly born).}

\item{GRR}{Genetic relative risk i.e. the constant multiplier on the hazard for disease for carriers compared to non-carriers.}
}
\description{
Simulates one pedigree.
Based on pedigree simulation methodology described by Nieuwoudt et al. (2018).
}
\details{
Pedigrees are simulated until one meets
the ascertainment requirements specified by the user, then it outputs only this pedigree.
The founder is a carrier of a dominant rare genetic variant. Individuals in the pedigree can
have children, experience disease onset, and die. Waiting times for these events
will be generated when an individual is created, and will be in the form of their age when the event happens.
These events will continue for individuals until the end of the ascertainment interval,
or censoring from death. Death is based on unaff_death_haz and disease onset on dis_haz, and both
can only happen once per individual. A total number of children will be generated for an individual
and their births will be uniformly distributed between the ages of 18 and 45. If an individual has children,
they will be generated a non-carrier partner of the opposite sex who is over 18 and is within 5 years
of the age of the individual. The partner will not have children waiting times, but
will have disease onset and death waiting times. Children of carriers will have a 50\%
probability of being a carrier. Children may not be born if the death waiting time
is earlier than the child's birth, or there is censoring from the end of the ascertainment
interval. After disease onset, a new waiting time will be generated using aff_death_haz.

A proband is a carrier who has disease onset within the user specified ascertainment
interval. For the pedigree to be ascertained, this proband must exist, and there must
also be at least the number of user specified affected (diseased) carriers throughout
the study time. This is between the time the founder is born and the end of the ascertainment
interval. The founder is born uniformly within an interval specified by the user.

A Gamma frailty term with a mean of 1 can be included to capture
between family heterogeneity in the general hazard of disease.
}
\examples{
# Load example hazard datasets
data("dis_haz")
data("aff_death_haz")
data("unaff_death_haz")

# Simulate a single pedigree
ped <- pedigree(aff = 2, frailty = 1, frail_var = 4.519, dis_haz = dis_haz,
unaff_death_haz = unaff_death_haz, aff_death_haz = aff_death_haz,
a_start = 2005, a_end = 2025, f_start = 1885, f_end = 1935, GRR = 19)

# View pedigree dataset
ped

}
\references{
Nieuwoudt, C., Jones, S. J., Brooks-Wilson, A., and Graham, J. (2018). Simulating pedigrees
ascertained for multiple disease-affected relatives. Source Code for Biology and Medicine,
13(2). https://doi.org/10.1186/s13029-018-0069-6
}
\seealso{
\code{\link[=multiPed]{multiPed()}}
}
