% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unaff_death_haz}
\alias{unaff_death_haz}
\title{Age-specific hazard of death in the general population.}
\format{
A numeric vector of length 100.
}
\source{
Xu, J., Murphy, S. L., Kochanek, K., and Arias, E. (2025). Deaths: Final data for 2022.
National Vital Statistics Reports Vol. 74, No. 4, National Center for Health Statistics.
}
\usage{
unaff_death_haz
}
\description{
A numeric vector containing the age-specific hazard rates of death
in the general population. The rates are expressed per person-year.
}
\keyword{datasets}
